bigCircles added

+++ haskell02.hs
@@ -16,5 +16,6 @@ itemize str = map (\str -> "<li>" ++ str ++ "</li>") str
 areaMaior :: Float -> Float -> Bool
 areaMaior num r = (r^2 * pi) > num
 
-bigCircles :: Float -> [Int] -> [Float]
-bigCircles num lista = filter (areaMaior num lista)
+bigCircles :: Float -> [Float] -> [Float]
+bigCircles num lista = filter (areaMaior num) lista
+
